Calculator progres

+++ ATC/ATC/ITrack.cs
@@ -10,8 +10,8 @@ namespace ATC
         int _xCord { get; }
         int _yCord { get; }
         int _alt { get; }
-        int _velocity { get; }
-        int _course { get; }
+        double _velocity { get; }
+        double _course { get; }
         DateTime _timestamp { get; }
     }
 }
